Bitcoin has been an altogether unpleasant experience
"Don't knock it until you try it. It's going to change the world. To the
moon!"
I tried it, and it's a worse version of the modern banking system, with
the only difference being different people are making the money.
My world is not changed, but vast amounts of other's worlds are
significantly poorer.
I'm not richer, and the people who are haven't contributed anything to
the world beyond social engineering.
The economy, like moderation and censorship, is an inherently social
problem. If it has a solution, it's one that's primarily social in
nature.
I'm rather sick of second-rate wannabe technocrats promising us
solutions that don't, in fact, solve anything.
But I gave them a fair shake, and it turned out to be everything I
already find unpleasant about the existing financial system, except
worse. Do yourself a favor if you're reading this, and don't.